% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{calculate_habitat_categories_intersect}
\alias{calculate_habitat_categories_intersect}
\title{Intersect habitat suitability feature layers with other categorical feature layer (e.g. Fire).}
\usage{
calculate_habitat_categories_intersect(
  taxa_list,
  targ_maxent_table,
  target_path,
  output_path,
  thresh_path,
  category_layer,
  numeric_col,
  template_raster,
  poly_path,
  epsg
)
}
\arguments{
\item{taxa_list}{Character string - The taxa to run maxent predictions for}

\item{targ_maxent_table}{data frame - table of maxent results for target taxa}

\item{target_path}{Character string - The file path containing the existing maxent models}

\item{output_path}{Character string - The file path to save the function output}

\item{thresh_path}{Character string - The file path to save the function output}

\item{category_layer}{Feature layer    - The file containing categorical features (e.g. fire intensity)}

\item{numeric_col}{Character string - The column for the category layer}

\item{template_raster}{Raster::raster - template raster with study extent and resolution}

\item{poly_path}{Character string - file path to feature polygon layer}

\item{epsg}{Numeric - ERSP code of coord ref system to be translated into WKT format}
}
\description{
Takes a habitat suitability layer, and intersects it with a categorical featurelayer.
}
