% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{calculate_taxa_habitat}
\alias{calculate_taxa_habitat}
\title{Intersect habitat suitability rasters with Fire layers.}
\usage{
calculate_taxa_habitat(
  taxa_list,
  targ_maxent_table,
  host_maxent_table,
  target_path,
  output_path,
  intersect_path,
  raster_pattern,
  fire_raster,
  cell_size,
  fire_thresh,
  write_rasters,
  country_shp,
  country_prj
)
}
\arguments{
\item{taxa_list}{Character string - The species to run maxent predictions for}

\item{targ_maxent_table}{Data frame       - A table of maxent results to be used for mapping}

\item{target_path}{Character string - The file path containing the existing maxent models}

\item{intersect_path}{Character string - The file path containing the intersecting rasters}

\item{raster_pattern}{Character string - The pattern to look for of Invertebrate rasters}

\item{cell_size}{Numeric          - Cell size to resample output}

\item{write_rasters}{Logical          - Save rasters (T/F)?}

\item{country_shp}{Character string - Shapefile name that has already been read into R (e.g. in the Package)}

\item{country_prj}{Character string - Name of projection}
}
\description{
Takes a habitat suitability layer, and intersects it with a fire suitability layer.
}
