% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{calculate_taxa_habitat_fire_features}
\alias{calculate_taxa_habitat_fire_features}
\title{Intersect habitat suitability features with Fire layers.}
\usage{
calculate_taxa_habitat_fire_features(
  taxa_list,
  analysis_df,
  taxa_level,
  targ_maxent_table,
  target_path,
  output_path,
  thresh_path,
  intersect_name,
  intersect_path,
  intersect_patt,
  main_int_layer,
  second_int_layer,
  template_raster,
  poly_path,
  epsg
)
}
\arguments{
\item{taxa_list}{Character string - The taxa to run maxent predictions for}

\item{analysis_df}{spdf - The spatial table of maxent records for all taxa}

\item{taxa_level}{Character string - The taxonomic level of analysis}

\item{targ_maxent_table}{data frame - table of maxent results for target taxa}

\item{target_path}{Character string - The file path containing the existing maxent models}

\item{output_path}{Character string - The file path containing the intersecting layers}

\item{thresh_path}{Character string - The file path containing the thresh-holded layers}

\item{intersect_name}{Character string - The file name for the output intersecting layers}

\item{intersect_path}{Character string - The file path containing the intersected layers}

\item{intersect_patt}{Character string - The pattern for the intersected layers}

\item{main_int_layer}{Simple features polygon - The main layer to intersect with the habitat layer (e.g. fire)}

\item{second_int_layer}{Simple features polygon - The 2nd layer to intersect with the habitat layer (e.g. Veg)}

\item{template_raster}{Raster::raster - Grid with the analysis extent and resolution}

\item{poly_path}{Character string - file path to feature polygon layer}

\item{epsg}{Numeric - ERSP code of coord ref system to be translated into WKT format}
}
\description{
Takes a habitat suitability layer, and intersects it with a fire suitability layer.
}
