% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{calculate_taxa_habitat_host_features}
\alias{calculate_taxa_habitat_host_features}
\title{Intersect habitat suitability feature layers with other feature layer (e.g. Fire).}
\usage{
calculate_taxa_habitat_host_features(
  taxa_list,
  targ_maxent_table,
  host_maxent_table,
  target_path,
  output_path,
  host_path,
  intersect_path,
  intersect_patt,
  int_cols,
  thresh_patt,
  main_int_layer,
  second_int_layer,
  template_raster,
  poly_path,
  epsg
)
}
\arguments{
\item{taxa_list}{Character string - The taxa to run maxent predictions for}

\item{targ_maxent_table}{data frame - table of maxent results for target taxa}

\item{host_maxent_table}{data frame - table of maxent results for host taxa}

\item{target_path}{Character string - The file path containing the existing maxent models}

\item{output_path}{Character string - The file path to save the function output}

\item{host_path}{Character string - The file path to save the function output}

\item{intersect_path}{Character string - The file path containing the intersecting rasters}

\item{intersect_patt}{Character string - The pattern for the intersect files}

\item{int_cols}{Character string - List of intersect columns}

\item{thresh_patt}{Character string - The pattern for the threshold files}

\item{main_int_layer}{Character string - The pattern for the intersect files}

\item{second_int_layer}{Character string - The pattern for the intersect files}

\item{template_raster}{Raster::raster - template raster with study extent and resolution}

\item{poly_path}{Character string - file path to feature polygon layer}

\item{epsg}{Numeric - ERSP code of coord ref system to be translated into WKT format}
}
\description{
Takes a habitat suitability layer, and intersects it with a fire suitability layer.
}
