% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{calculate_taxa_habitat_host_rasters}
\alias{calculate_taxa_habitat_host_rasters}
\title{Intersect habitat suitability raster layers with other raster layer (e.g. Fire).}
\usage{
calculate_taxa_habitat_host_rasters(
  taxa_list,
  targ_maxent_table,
  host_maxent_table,
  target_path,
  output_path,
  intersect_path,
  raster_pattern,
  fire_raster,
  cell_size,
  fire_thresh,
  write_rasters,
  poly_path,
  epsg
)
}
\arguments{
\item{taxa_list}{Character string - The taxa to run maxent predictions for}

\item{targ_maxent_table}{Data frame       - A table of maxent results to be used for mapping}

\item{host_maxent_table}{data frame - table of maxent results for host taxa}

\item{target_path}{Character string - The file path containing the existing maxent models}

\item{intersect_path}{Character string - The file path containing the intersecting rasters}

\item{raster_pattern}{Character string - The pattern to look for of Invertebrate rasters}

\item{cell_size}{Numeric          - Cell size to resample output}

\item{write_rasters}{Logical          - Save rasters (T/F)?}

\item{poly_path}{Character string - file path to feature polygon layer}

\item{epsg}{Numeric - ERSP code of coord ref system to be translated into WKT format}
}
\description{
Takes a habitat suitability layer, and intersects it with a fire suitability layer.
}
