% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{combine_ala_records}
\alias{combine_ala_records}
\title{Combine ALA Records for different taxa!}
\usage{
combine_ala_records(
  taxa_list,
  records_path,
  records_extension,
  record_type,
  keep_cols,
  year_filt,
  year,
  year_lim,
  unique_cells,
  world_raster
)
}
\arguments{
\item{taxa_list}{Character Vector - List of taxa already downloaded}

\item{records_path}{File path for downloaded taxa}

\item{records_extension}{Which R file type? RDS or RDA}

\item{record_type}{Adds a column to the data frame for the data source, EG ALA}

\item{keep_cols}{The columns we want to keep - a character list created by you}

\item{year_filt}{Character - filter out records based on a year?}

\item{year_lim}{Numeric - Records to remove}

\item{unique_cells}{Logical - take only one record per cell?}

\item{world_raster}{An Raster file of the enviro conditions used (assumed to be global)}
}
\description{
Combines all occurrence files from ALA into one table.
It assumes that all the files come from the previous downloads function.
Although you can download all the records for in one go, this is better for
Doing small runs of species, or where you want to re-run them constantly
}
