% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{combine_records_extract}
\alias{combine_records_extract}
\title{Extract environmental records for occurrence data.}
\usage{
combine_records_extract(
  ala_df,
  add_sites,
  site_df,
  add_site,
  filter_taxo,
  taxa_list,
  taxa_level,
  template_raster,
  thin_records,
  world_raster,
  prj,
  complete_var,
  raster_divide,
  env_variables,
  save_data,
  data_path,
  save_run
)
}
\arguments{
\item{ala_df}{Data frame of ALA records}

\item{site_df}{Data frame of site records (only used if you have site data, e.g. I-naturalist)}

\item{taxa_list}{List of taxa analyzed, used to cut the dataframe down}

\item{taxa_level}{What taxonomic level to analyze at?}

\item{template_raster}{A global R Raster used to thin records to 1 record per 1km grid cell}

\item{thin_records}{Do you want to thin the records out? If so, it will be 1 record per 1km*1km grid cell}

\item{world_raster}{An global R Raster of the enviro conditions used to extract values for all records}

\item{prj}{The projection system used. Currently, needs to be WGS84}

\item{save_data}{Do you want to save the data frame?}

\item{data_path}{The file path used for saving the data frame}

\item{save_run}{A run name to append to the data frame (e.g. bat taxa, etc.). Useful for multiple runs.}

\item{biocl_vars}{The variables used - eg the standard bioclim names (https://www.worldclim.org/).}

\item{env_vars}{The actual variable names (e.g. bio1 = rainfall, etc.) Only needed for worldlcim}

\item{worldclim_divide}{Are you using worldclim stored as long intergers? If so, divide by 10.}
}
\description{
This function combines occurrence files from ALA and GBIF into one table, and extracts enviro values.
It assumes that both files come from the previous GBIF/ALA combine function.
}
