% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAXENT_FUNCTIONS.R
\name{compile_sdm_results}
\alias{compile_sdm_results}
\title{Compile SDM results}
\usage{
compile_sdm_results(
  taxa_list,
  results_dir,
  save_data,
  data_path,
  sdm_path,
  save_run
)
}
\arguments{
\item{taxa_list}{Character string - the taxa to run maxent models for}

\item{results_dir}{Character string - The file path used for saving the maxent output}

\item{save_data}{Logical or character - do you want to save the data frame?}

\item{data_path}{Character string - The file path used for saving the data frame}

\item{sdm_path}{Character string - The file path where the maxent models are stored.}

\item{save_run}{Character string - run name to append to the data frame, useful for multiple runs.}
}
\description{
This function extracts the SDM results from the folders.
}
