% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{diverge0}
\alias{diverge0}
\title{Plot a rasterVis::levelplot}
\usage{
diverge0(p, ramp)
}
\arguments{
\item{p}{A trellis object resulting from rasterVis::levelplot}

\item{ramp}{Character string - The name of an RColorBrewer palette (as character), a character}
}
\description{
Plot a rasterVis::levelplot with a colour ramp diverging around zero
Adapted from a gist by John Baumgartner for the (https://gist.github.com/johnbaums?direction=desc&sort=updated).
}
