% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{format_ala_dump}
\alias{format_ala_dump}
\title{Format ALA Records from a big data download}
\usage{
format_ala_dump(
  ALA_table,
  record_type,
  keep_cols,
  year_filt,
  year_lim,
  unique_cells,
  world_raster
)
}
\arguments{
\item{ALA_table}{Character Vector - List of taxa already downloaded}

\item{record_type}{Adds a column to the data frame for the data source, EG ALA}

\item{keep_cols}{The columns we want to keep - a character list created by you}

\item{year_filt}{Logical - filter out records based on a year?}

\item{year_lim}{Numeric - Records to remove}

\item{unique_cells}{Logical - take only one record per cell?}

\item{world_raster}{An Raster file of the enviro conditions used (assumed to be global)}
}
\description{
Combines all occurrence files from ALA into one table.
It assumes that all the input table comes from the ALA data dump.
}
