% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{gdal_resample}
\alias{gdal_resample}
\title{Resampling a Raster* object via rGDAL.}
\usage{
gdal_resample(
  rast,
  rast_base,
  method,
  temp_dir,
  write_rasters,
  output_path,
  output_file
)
}
\arguments{
\item{rast}{Raster* object to be resampled}

\item{rast_base}{Raster* object with parameters that r
should be resampled to.}

\item{method}{Character. GDAL resampling_method
("near"|"bilinear"|"cubic"|"cubicspline"|
 "lanczos"|"average"|"mode"|"max"|"min"|
 "med"|"q1"|"q3")}

\item{temp_dir}{Character.}

\item{write_rasters}{Character.}

\item{output_path}{Character.}

\item{output_file}{Character.}
}
\description{
Resampling a Raster* object via rGDAL.
}
