% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{habitat_threshold}
\alias{habitat_threshold}
\title{Threshold current habitat suitability rasters.}
\usage{
habitat_threshold(
  taxa_list,
  maxent_table,
  maxent_path,
  cell_factor,
  country_shp,
  country_prj
)
}
\arguments{
\item{taxa_list}{Character string - The species to run maxent predictions for}

\item{maxent_table}{Data frame       - A table of maxent results to be used for mapping}

\item{maxent_path}{Character string - The file path containing the existing maxent models}

\item{cell_factor}{Numeric          - Cell size to resample output}

\item{country_shp}{Character string - Shapefile name that has already been read into R (e.g. in the Package)}

\item{country_prj}{Character string - Name of projection}

\item{write_rasters}{Logical          - Save rasters (T/F)?}
}
\description{
Takes a habitat suitability layer, and creates a binary suitability layer (0, 1) using a threshold value.
}
