% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{project_maxent_current_grids_mess}
\alias{project_maxent_current_grids_mess}
\title{This function takes the maxent models created by the 'fit_maxent_targ_bg_back_sel' function,
and projects the model across geographic space - currently just for Australia.
It uses the rmaxent package https://github.com/johnbaums/rmaxent
It assumes that the maxent models were generated by the 'fit_maxent_targ_bg_back_sel'}
\usage{
project_maxent_current_grids_mess(
  country_shp,
  country_prj,
  save_novel_poly,
  local_prj,
  species_list,
  maxent_path,
  climate_path,
  grid_names,
  current_grids,
  create_mess
)
}
\arguments{
\item{country_shp}{SpatialPolygonsDataFrame - Spdf of the country for mapping maxent results (e.g. Australia)}

\item{country_prj}{CRS object  - Local projection for mapping maxent results}

\item{save_novel_poly}{Character string - Save the novel areas as shapefiles?}

\item{local_prj}{CRS object  - Local projection for mapping maxent results}

\item{species_list}{Character string - The species to run maxent predictions for}

\item{maxent_path}{Character string - The file path containin the existing maxent models}

\item{climate_path}{Character string - The file path where the climate data is saved}

\item{grid_names}{Character string - Vector of enviro conditions that you want to include}

\item{current_grids}{Character string - Vector of current enviro conditions that you want to include}

\item{create_mess}{Logical - Create mess maps of the predictions (T/F)?}
}
\description{
This function takes the maxent models created by the 'fit_maxent_targ_bg_back_sel' function,
and projects the model across geographic space - currently just for Australia.
It uses the rmaxent package https://github.com/johnbaums/rmaxent
It assumes that the maxent models were generated by the 'fit_maxent_targ_bg_back_sel'
}
