% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PLOTTING_FUNCTIONS.R
\name{single_bar_order_factor}
\alias{single_bar_order_factor}
\title{Project current maxent models across geographic space}
\usage{
single_bar_order_factor(
  df,
  title,
  xvar,
  yvar,
  ylab,
  xlab,
  ymin,
  ymax,
  axis_multiplier,
  col_palette,
  tsize,
  caption,
  xsize,
  ysize,
  ycol,
  mar,
  lab_size,
  bar_width,
  capt_size
)
}
\arguments{
\item{taxa_list}{Character string - The taxa to run maxent predictions for}

\item{maxent_path}{Character string - The file path containing the existing maxent models}

\item{output_path}{Character string - The file path where the polygon data will be saved}

\item{current_grids}{Character string - Vector of current enviro conditions that you want to include}

\item{save_novel_poly}{Character string - Save the novel areas as polygons?}

\item{create_mess}{Logical - Create mess maps of the predictions (T/F)?}

\item{mess_layers}{Logical - Save mess maps of the each layer as .png (T/F)?}

\item{poly_path}{Character string - file path to feature polygon layer.}

\item{epsg}{Numeric - ERSP code of coord ref system to be translated into WKT format}
}
\description{
This function takes the maxent models created by the 'fit_maxent_targ_bg_back_sel' function,
and projects the model across geographic space - currently just for Australia.
}
\details{
It uses the rmaxent package https://github.com/johnbaums/rmaxent
}
