% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{split_shp}
\alias{split_shp}
\title{Split SDM table}
\usage{
split_shp(taxa_list, sdm_df, data_path)
}
\arguments{
\item{taxa_list}{Character string - List of species}

\item{data_path}{Character string - The file path used for saving the data frame}

\item{sdm_}{Path of taxa records, with spatial outlier T/F flag for each record}
}
\description{
'This function takes a shapefile of all taxa records,
and splits in into shp files for each taxa
}
