% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{taxa_records_habitat_features_intersect}
\alias{taxa_records_habitat_features_intersect}
\title{Taxa records intersect}
\usage{
taxa_records_habitat_features_intersect(
  analysis_df,
  taxa_list,
  taxa_level,
  habitat_poly,
  output_path,
  buffer,
  raster_convert,
  int_cols,
  save_shp,
  save_png,
  epsg,
  poly_path
)
}
\arguments{
\item{analysis_df}{SpatialPolygonsDataFrame - Spdf of all the taxa analysed}

\item{taxa_list}{Character string - The taxa to run maxent predictions for}

\item{taxa_level}{Character string - the taxnomic level to run maxent models for}

\item{buffer}{Numeric          - Distance by which to buffer the points (metres using a projected system)}

\item{raster_convert}{Logical          - Convert to raster?}

\item{int_cols}{Character string - list of columns to keep from records * layer intersect}

\item{save_shp}{Logical          - Save as .shp? Geopackage is much better.}

\item{save_png}{Logical          - Save as .png?}

\item{epsg}{Numeric - ERSP code of coord ref system to be translated into WKT format}

\item{poly_path}{Character string - file path to feature polygon layer}

\item{habitat_raster}{Character string - The habitat raster which has already been read in}

\item{country_shp}{Character string - Shapefile name that has already been read into R (e.g. in the Package)}
}
\description{
Take a table of taxa records, and intersect the records for each taxon with a feature 
layer (e.g. Vegetation).
}
