% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{taxa_records_habitat_intersect}
\alias{taxa_records_habitat_intersect}
\title{Taxa records intersect}
\usage{
taxa_records_habitat_intersect(
  analysis_df,
  taxa_list,
  taxa_level,
  habitat_poly,
  output_path,
  buffer
)
}
\arguments{
\item{analysis_df}{SpatialPolygonsDataFrame - Spdf of all the taxa analysed}

\item{taxa_list}{Character string - The species to run maxent predictions for}

\item{taxa_level}{Character string - the taxnomic level to run maxent models for}

\item{buffer}{Numeric          - Distance by which to buffer the points (metres using a projected system)}

\item{habitat_raster}{Character string - The habitat raster which has already been read in}

\item{country_shp}{Character string - Shapefile name that has already been read into R (e.g. in the Package)}

\item{write_rasters}{Logical          - Save rasters (T/F)?}
}
\description{
Take a table of taxa records, and intersect the records for each taxa with a shapefile  
of habitat (e.g. Vegetation).
}
