% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAXENT_FUNCTIONS.R
\name{var_importance}
\alias{var_importance}
\title{Variable importance}
\usage{
var_importance(mod)
}
\arguments{
\item{mod}{object - maxent model object}
}
\description{
This function calculates variables importance.
}
