% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{completeFun}
\alias{completeFun}
\title{Complete data frame}
\usage{
completeFun(data, desiredCols)
}
\arguments{
\item{data}{Data.frame of taxa records}

\item{desiredCols}{Character string of columns to search for NA values EG c('rainfall', 'temp'), etc.}
}
\value{
A df with NA records removed
}
\description{
Remove taxa records with NA enviro (i.e. Raster) values - records outside the exetent of rasters
}
