% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{download_ALA_all_genera}
\alias{download_ALA_all_genera}
\title{Download ALA genus occurrences.}
\usage{
download_ALA_all_genera(
  genera_list,
  your_email,
  download_path,
  ala_temp_dir,
  download_limit,
  quality_cols
)
}
\arguments{
\item{download_path}{Character string - File path for genus downloads}

\item{download_limit}{Numeric - How many records can be downloaded at one time? Set by server}

\item{quality_cols}{Character - quality ALA columns, eg spatial accuracy}

\item{species_list}{Character vector - List of genus binomials to download}

\item{extra_cols}{Character - extra ALA columns, eg environmental vatriables}
}
\value{
Data frame of all site records, with global enviro conditions for each record location (i.e. lat/lon)
}
\description{
This function downloads genus occurrence files from ALA (https://www.ala.org.au/).
It assumes that the genus list supplied is taxonomically correct.
It downloads the genus to fiel without returning anything to the global environment
}
