% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_PROCESSOR_FUNCTIONS.R
\name{prepare_sdm_table}
\alias{prepare_sdm_table}
\title{Prepare SDM table.}
\usage{
prepare_sdm_table(
  coord_df,
  taxa_list,
  site_flag,
  occ_flag,
  site_records,
  BG_points,
  sdm_table_vars,
  save_run,
  read_background,
  background_points,
  save_data,
  data_path,
  sp_country_prj
)
}
\arguments{
\item{coord_df}{Data.frame. DF of all taxa records returned by the coord_clean_records function}

\item{taxa_list}{Character string - the taxa analysed}

\item{BG_points}{Logical - Do we want to include a dataframe of background points? Otherwise, BG points taken from taxa not modelled}

\item{sdm_table_vars}{Character string - The enviro vars to be included in taxa models}

\item{save_run}{Character string - append a run name to the output (e.g. 'bat_taxa')}

\item{read_background}{Logical - Read in an additional dataframe of background points (T/F)?}

\item{background_points}{Data.frame. DF of extra taxa records}

\item{save_data}{Logical - do you want to save the data frame?}

\item{data_path}{Character string - The file path used for saving the data frame}

\item{save_shp}{Logical - Save a shapefile of the SDM table (T/F)?}

\item{project_path}{Paht of taxa records, with spatial outlier T/F flag for each record}
}
\description{
'This function takes a data frame of all taxa records,
And prepares a table in the 'taxa with data' (swd) format for modelling uses the Maxent algorithm.
It assumes that the input df is that returned by the coord_clean_records function'
}
