% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAPPING_FUNCTIONS.R
\name{project_maxent_current_grids_mess}
\alias{project_maxent_current_grids_mess}
\title{Project current maxent models across geographic space}
\usage{
project_maxent_current_grids_mess(
  taxa_list,
  maxent_path,
  current_grids,
  create_mess,
  save_novel_poly,
  output_path,
  poly_path,
  epsg
)
}
\arguments{
\item{taxa_list}{Character string - The taxa to run maxent predictions for}

\item{maxent_path}{Character string - The file path containing the existing maxent models}

\item{current_grids}{Character string - Vector of current enviro conditions that you want to include}

\item{create_mess}{Logical - Create mess maps of the predictions (T/F)?}

\item{save_novel_poly}{Character string - Save the novel areas as polygons?}

\item{output_path}{Character string - The file path where the polygon data will be saved}

\item{poly_path}{Character string - file path to feature polygon layer.}

\item{epsg}{Numeric - ERSP code of coord ref system to be translated into WKT format}
}
\description{
This function takes the maxent models created by the 'fit_maxent_targ_bg_back_sel' function,
and projects the model across geographic space - currently just for Australia.
}
\details{
It uses the rmaxent package https://github.com/johnbaums/rmaxent
}
