% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM_GEN_MAXENT_FUNCTIONS.R
\name{run_sdm_analysis_no_crop}
\alias{run_sdm_analysis_no_crop}
\title{Run SDM analyses for 'species with data' format.}
\usage{
run_sdm_analysis_no_crop(
  taxa_list,
  taxa_level,
  sdm_df,
  sdm_predictors,
  maxent_dir,
  bs_dir,
  backwards_sel,
  template_raster,
  cor_thr,
  pct_thr,
  k_thr,
  min_n,
  max_bg_size,
  background_buffer_width,
  features,
  feat_save,
  replicates,
  responsecurves,
  poly_path,
  epsg
)
}
\arguments{
\item{taxa_list}{Character string - the taxa to run maxent models for}

\item{taxa_level}{Character string - the taxnomic level to run maxent models for}

\item{sdm_df}{SpatialPointsDataFrame. Spdf of all taxa records returned by the 'prepare_sdm_table' function}

\item{sdm_predictors}{Character string - Vector of enviro conditions that you want to include}

\item{maxent_dir}{Character string - The file path used for saving the maxent output}

\item{bs_dir}{Character string - The file path used for saving the backwards selection maxent output}

\item{backwards_sel}{Logical - Run backwards selection using the maxent models (T/F)?}

\item{template_raster}{RasterLayer - Empty raster with analysis extent (global), res (1km) and projection (Mollweide, EPSG 54009)}

\item{cor_thr}{Numeric - The max allowable pairwise correlation between predictor variables}

\item{pct_thr}{Numeric - The min allowable percent variable contribution}

\item{k_thr}{Numeric - The min number of variables to be kept in the model}

\item{min_n}{Numeric - The min number of records for running maxent}

\item{max_bg_size}{Numeric - The max number of background points to keep}

\item{background_buffer_width}{Numeric - The max distance (km) from occ points that BG points should be selected?}

\item{features}{Character string - Which features should be used? (e.g. linear, product, quadratic 'lpq')}

\item{feat_save}{Logical - Save features of the occ and bg data (T/F)?}

\item{replicates}{Numeric - The number of replicates to use}

\item{responsecurves}{Logical - Save response curves of the maxent models (T/F)?}

\item{poly_path}{Character string - file path to feature polygon layer}

\item{epsg}{Numeric - ERSP code of coord ref system to be translated into WKT format}
}
\description{
This function takes a data frame of all taxa records,
and runs a specialised maxent analysis for each taxa.
It uses the dismo package https://github.com/johnbaums/rmaxent
}
